Comment out setting with period
Due to how hocon expects keys to be defined in the setting metaparam, we
cannot properly have "dujour.core" as a string in the config file
because it thinks dujour and core are both keys. This commit just
comments out that value in the config file since the default value laid
down by the package is good enough.